requirements.txt
boto3==1.26.150
psycopg2-binary==2.9.6
python-dotenv==1.0.0